Slack_bot/requirements.txt
requests
dnspython
pymongo
SQLalchemy
pandas 
slackclient
nest_asyncio
psycopg2-binary